support 9.7 tests
