+++ .github/workflows/docker-tests-9.7.yml
@@ -59,23 +59,23 @@ jobs:
           - 'VERSION=9.7 GODOG_FEATURE=free_space.feature make test'
           - 'VERSION=9.7 GODOG_FEATURE=host_discovery.feature make test'
           - 'VERSION=9.7 GODOG_FEATURE=host_management.feature make test'
+          - 'VERSION=9.7 GODOG_FEATURE=light_maintenance_mode.97.feature make test'
           - 'VERSION=9.7 GODOG_FEATURE=maintenance.97.feature make test'
+          - 'VERSION=9.7 GODOG_FEATURE=manager_switchover.feature make test'
           - 'VERSION=9.7 GODOG_FEATURE=offline_mode.84.feature make test'
           - 'VERSION=9.7 GODOG_FEATURE=offline_mode_az_limit.feature make test'
+          - 'VERSION=9.7 GODOG_FEATURE=optimization.feature make test'
           - 'VERSION=9.7 GODOG_FEATURE=priority.feature make test'
           - 'VERSION=9.7 GODOG_FEATURE=readonly_filesystem.feature make test'
           - 'VERSION=9.7 GODOG_FEATURE=recovery.97.feature make test'
           - 'VERSION=9.7 GODOG_FEATURE=repair.97.feature make test'
-          - 'VERSION=9.7 GODOG_FEATURE=semisync_source_replica.97.feature make test'
           - 'VERSION=9.7 GODOG_FEATURE=repl_mon.feature make test'
+          - 'VERSION=9.7 GODOG_FEATURE=semisync_source_replica.97.feature make test'
           - 'VERSION=9.7 GODOG_FEATURE=statefile.feature make test'
           - 'VERSION=9.7 GODOG_FEATURE=switchover_from.97.feature make test'
           - 'VERSION=9.7 GODOG_FEATURE=switchover_to.97.feature make test'
           - 'VERSION=9.7 GODOG_FEATURE=zk_failure.97.feature make test'
           - 'VERSION=9.7 GODOG_FEATURE=zk_maintenance.feature make test'
-          - 'VERSION=9.7 GODOG_FEATURE=manager_switchover.feature make test'
-          - 'VERSION=9.7 GODOG_FEATURE=optimization.feature make test'
-          - 'VERSION=9.7 GODOG_FEATURE=light_maintenance_mode.97.feature make test'
       fail-fast: false
 
     steps:
