+++ tests/images/docker-compose.yaml
@@ -96,7 +96,6 @@ services:
       MYSYNC_STREAM_FROM_REASONABLE_LAG:
       MYSYNC_RESETUP_CRASHED_HOSTS:
       MYSYNC_REPLICATION_REPAIR_COOLDOWN:
-      MYSYNC_REPLICATION_REPAIR_MAX_ATTEMPTS:
       MYSYNC_REPLICATION_REPAIR_AGGRESSIVE_MODE:
       MYSYNC_SET_RO_TIMEOUT:
       MYSYNC_REPLICATION_LAG_QUERY:
@@ -155,7 +154,6 @@ services:
       MYSYNC_STREAM_FROM_REASONABLE_LAG:
       MYSYNC_RESETUP_CRASHED_HOSTS:
       MYSYNC_REPLICATION_REPAIR_COOLDOWN:
-      MYSYNC_REPLICATION_REPAIR_MAX_ATTEMPTS:
       MYSYNC_REPLICATION_REPAIR_AGGRESSIVE_MODE:
       MYSYNC_SET_RO_TIMEOUT:
       MYSYNC_REPLICATION_LAG_QUERY:
@@ -208,7 +206,6 @@ services:
       MYSYNC_STREAM_FROM_REASONABLE_LAG:
       MYSYNC_RESETUP_CRASHED_HOSTS:
       MYSYNC_REPLICATION_REPAIR_COOLDOWN:
-      MYSYNC_REPLICATION_REPAIR_MAX_ATTEMPTS:
       MYSYNC_REPLICATION_REPAIR_AGGRESSIVE_MODE:
       MYSYNC_SET_RO_TIMEOUT:
       MYSYNC_REPLICATION_LAG_QUERY:
